refactor: extract the Calendar tab, which now owns its own state
nav/CalendarTab.tsx takes the month/selected-day state, both queries
(user-profile, calendar-events), the day-list auto-scroll and EventRow. Props
are just { recordings, isPersonalRoom }; selection comes from context, as it
already does in ActionsTab/ActionsToolbar.

The panel mounts it only while the tab is showing, so the enabled: tab ===
"calendar" gate becomes conditional mounting. Two consequences, both intended
and both stated in the release notes: the queries no longer run from other
tabs, and the month resets to today on re-entry.

Two memo fixes while the code was open: selectedItems was recomputed every
render, and `events` allocated a fresh [] per render in a shared room, which
defeated the eventKeys memo on the very next line.

Red test first (the import did not resolve), then mutation-checked: making the
month survive unmount turns the reset test red and nothing else, so it is
asserting the behaviour rather than passing by construction.

Panel 720 -> 574. The panel's own test file is still byte-identical.
