+++ apps/web/src/components/RecordingsPanel.tsx
@@ -1,7 +1,5 @@
 import { useEffect, useMemo, useRef, useState } from "react";
-import { NavLink } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import type { TFunction } from "i18next";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { api, apiErrorMessage } from "../lib/api";
 import { usePanelTab, setPanelTab } from "../lib/panelTab";
@@ -26,8 +24,7 @@ import TabStrip from "./nav/TabStrip";
 import ListToolbar from "./nav/ListToolbar";
 import UploadStatusList from "./nav/UploadStatusList";
 import GroupSelectCheckbox from "./nav/GroupSelectCheckbox";
-import MonthCalendar from "./MonthCalendar";
-import { recordingDayKeys, dayKey, eventDayKeys, visibleGridRange, dayItems } from "../lib/calendar";
+import CalendarTab from "./nav/CalendarTab";
 import { useUpload } from "../lib/uploadContext";
 import { distinctActors, filterActions } from "../lib/actionsView";
 import { recordingsForTags, topTagsByCount } from "../lib/tagCloud";
@@ -36,7 +33,7 @@ import ActionsTab from "./ActionsTab";
 import EditActionModal from "./EditActionModal";
 import TagCloud from "./TagCloud";
 import TagCloudModal from "./TagCloudModal";
-import type { ActionListItem, CalendarEvent, RecordingSummary } from "../lib/types";
+import type { ActionListItem, RecordingSummary } from "../lib/types";
 import { RoomPermission } from "../lib/types";
 
 const TAG_LIMIT_KEY = "diariz.recordings.tagLimit";
@@ -57,7 +54,7 @@ function storedTagLimit(): number {
 /// The recordings list for the left panel, grouped into user sections (Ungrouped last).
 /// Selecting a row routes to /recordings/:id (middle panel).
 export default function RecordingsPanel() {
-  const { t, i18n } = useTranslation("workspace");
+  const { t } = useTranslation("workspace");
   const qc = useQueryClient();
   // The room being browsed (the switcher's current room). The recordings + folders lists are scoped to it;
   // the personal room keeps its folders/drag-drop, a shared room shows the recordings shared into it.
@@ -193,40 +190,6 @@ export default function RecordingsPanel() {
     [recordings, shownTags, selectedTag],
   );
 
-  const [month, setMonth] = useState(() => {
-    const now = new Date();
-    return { year: now.getFullYear(), month: now.getMonth() };
-  });
-  const [selectedDay, setSelectedDay] = useState<string | null>(() => dayKey(new Date()));
-  const dayKeys = useMemo(() => recordingDayKeys(recordings), [recordings]);
-
-  // Google Calendar overlay: fetch the visible month's events (only when the user has connected Calendar).
-  // Keyed by month, so navigating months auto-refetches; a short staleTime avoids refetch churn on focus.
-  const { data: profile } = useQuery({ queryKey: ["user-profile"], queryFn: api.getProfile });
-  const calendarConnected = profile?.googleCalendar === true;
-  const { data: calendarEvents = [], isFetching: eventsFetching } = useQuery({
-    queryKey: ["calendar-events", month.year, month.month],
-    queryFn: () => {
-      const { timeMin, timeMax } = visibleGridRange(month.year, month.month);
-      return api.getCalendarEvents(timeMin, timeMax);
-    },
-    // A shared room shows only its recordings on the calendar - no personal Google-event overlay.
-    enabled: calendarConnected && isPersonalRoom,
-    staleTime: 5 * 60_000,
-    retry: false,
-  });
-  // The Google overlay is personal-only. Force events empty in a shared room (the disabled query still holds
-  // the last personal-room data in cache - the key is room-agnostic - so gate the derived value, not just the fetch).
-  const showCalendarOverlay = calendarConnected && isPersonalRoom;
-  const events = isPersonalRoom ? calendarEvents : [];
-  const eventKeys = useMemo(() => eventDayKeys(events), [events]);
-  const selectedItems = selectedDay ? dayItems(recordings, events, selectedDay) : [];
-  function stepMonth(delta: number) {
-    setMonth((m) => {
-      const d = new Date(m.year, m.month + delta, 1);
-      return { year: d.getFullYear(), month: d.getMonth() };
-    });
-  }
   /// Apply a drag-and-drop: set the dragged recordings' group + order within the current room, then refresh.
   /// Order and folders are per-room, so this needs manage-contents (the personal-room owner always has it).
   ///
@@ -312,9 +275,8 @@ export default function RecordingsPanel() {
   const [dragging, setDragging] = useState(false);
   const dragDepth = useRef(0);
   // Native HTML5 DnD doesn't scroll the list while dragging near its edges, so a drop target outside the
-  // viewport is unreachable in a long tree. These auto-scroll the list / day-list during any drag.
+  // viewport is unreachable in a long tree. This auto-scrolls the list during any drag.
   const listScrollRef = useDragAutoScroll<HTMLDivElement>();
-  const dayScrollRef = useDragAutoScroll<HTMLDivElement>();
   function onFileDragEnter(e: React.DragEvent) {
     if (!dragHasFiles(e)) return;
     dragDepth.current += 1;
@@ -533,62 +495,7 @@ export default function RecordingsPanel() {
             )}
           </div>
         ) : tab === "calendar" ? (
-          // Calendar: the month grid stays fixed at the top; only the selected day's list scrolls.
-          // min-w-0 is essential: without it this flex child grows to the widest day-list row, which would
-          // stretch the grid-cols-7 month grid wider than the panel and make the calendar appear to resize
-          // when you pick a day with longer recording names.
-          <div className="flex min-h-0 min-w-0 flex-1 flex-col">
-            <div className="shrink-0 border-b dark:border-gray-800">
-              <MonthCalendar
-                year={month.year}
-                month={month.month}
-                daysWithRecordings={dayKeys}
-                daysWithEvents={showCalendarOverlay ? eventKeys : undefined}
-                selectedKey={selectedDay}
-                onSelect={setSelectedDay}
-                onPrev={() => stepMonth(-1)}
-                onNext={() => stepMonth(1)}
-              />
-              {showCalendarOverlay && (
-                <div className="flex items-center justify-end px-2 pb-1">
-                  <button
-                    type="button"
-                    onClick={() => qc.invalidateQueries({ queryKey: ["calendar-events", month.year, month.month] })}
-                    disabled={eventsFetching}
-                    className="text-[10px] text-gray-400 hover:text-gray-600 disabled:opacity-50 dark:text-gray-500 dark:hover:text-gray-300"
-                  >
-                    {eventsFetching ? t("calRefreshing") : t("calRefreshEvents")}
-                  </button>
-                </div>
-              )}
-            </div>
-            {/* Reserve the scrollbar gutter so toggling the day list's scrollbar never shifts its width. */}
-            <div ref={dayScrollRef} className="min-h-0 min-w-0 flex-1 overflow-y-auto [scrollbar-gutter:stable]">
-              {selectedItems.length === 0 ? (
-                <p className="p-4 text-sm text-gray-500 dark:text-gray-400">
-                  {showCalendarOverlay ? t("calDayEmpty") : t("calNoRecordings")}
-                </p>
-              ) : (
-                <ul className="divide-y dark:divide-gray-800">
-                  {selectedItems.map((it) =>
-                    it.type === "recording" ? (
-                      <RecordingRow
-                        key={it.recording.id}
-                        r={it.recording}
-                        indentClass="pl-3"
-                        selectMode={selection.selectMode}
-                        selected={selection.selectedIds.includes(it.recording.id)}
-                        onToggleSelect={() => selection.toggle(it.recording.id)}
-                        onDropBefore={() => {}}
-                      />
-                    ) : (
-                      <EventRow key={`ev-${it.event.id}`} event={it.event} locale={i18n.language} t={t} />
-                    ),
-                  )}
-                </ul>
-              )}
-            </div>
-          </div>
+          <CalendarTab recordings={recordings} isPersonalRoom={isPersonalRoom} />
         ) : tab === "actions" ? (
           // Actions: a flat, cross-transcript list with its own filter/select/complete toolbar above.
           <div className="min-h-0 min-w-0 flex-1 overflow-y-auto">
@@ -665,56 +572,3 @@ export default function RecordingsPanel() {
   );
 }
 
-/// A Google Calendar event row in the Calendar tab's merged day list — time range + title. Clicking the row
-/// opens the event preview (a meeting with no recording); the calendar glyph still links out to Google.
-/// Only unlinked events reach this row (a linked event is shown by its recording row, deduped in `dayItems`).
-/// Events from an external .ics feed (`calendarId` starting `ics:`) are display-only - they have no Google
-/// event to preview or link a recording to - so their row is a static (non-clickable) block, still coloured.
-function EventRow({ event, locale, t }: { event: CalendarEvent; locale: string; t: TFunction }) {
-  const fmt = new Intl.DateTimeFormat(locale, { hour: "2-digit", minute: "2-digit" });
-  const title = event.summary || t("calUntitledEvent");
-  const range = `${fmt.format(new Date(event.start))} – ${fmt.format(new Date(event.end))}`;
-  const isFeed = event.calendarId?.startsWith("ics:") ?? false;
-
-  const inner = (
-    <>
-      <svg
-        {...iconProps}
-        style={event.color ? { color: event.color } : undefined}
-        className={`mt-0.5 h-3.5 w-3.5 shrink-0 ${event.color ? "" : "text-green-600 dark:text-green-400"}`}
-        aria-label={t("calEventLabel")}
-      >
-        <rect x="3" y="4" width="18" height="18" rx="2" ry="2" />
-        <line x1="16" y1="2" x2="16" y2="6" />
-        <line x1="8" y1="2" x2="8" y2="6" />
-        <line x1="3" y1="10" x2="21" y2="10" />
-      </svg>
-      <div className="min-w-0 flex-1">
-        <div className="truncate text-gray-800 dark:text-gray-200">{title}</div>
-        <div className="flex items-center gap-1.5 text-xs text-gray-500 dark:text-gray-400">
-          <span className="tabular-nums">{range}</span>
-          {event.calendarName && <span className="truncate">· {event.calendarName}</span>}
-        </div>
-      </div>
-    </>
-  );
-
-  return (
-    <li>
-      {isFeed ? (
-        <div className="flex items-start gap-2 py-1.5 pl-3 pr-2 text-sm">{inner}</div>
-      ) : (
-        <NavLink
-          to={`/calendar-event/${encodeURIComponent(event.id)}`}
-          className={({ isActive }) =>
-            `flex items-start gap-2 py-1.5 pl-3 pr-2 text-sm ${
-              isActive ? "bg-blue-50 dark:bg-blue-900/30" : "hover:bg-gray-50 dark:hover:bg-gray-800"
-            }`
-          }
-        >
-          {inner}
-        </NavLink>
-      )}
-    </li>
-  );
-}
