+++ src/Diariz.Api/Diariz.Api.csproj
@@ -5,7 +5,7 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <!-- App version ÃƒÂ¢Ã¢â€šÂ¬Ã¢â‚¬Â keep in sync with /version.json (see CLAUDE.md versioning rule). -->
-    <Version>0.186.7</Version>
+    <Version>0.186.8</Version>
   </PropertyGroup>
 
   <ItemGroup>
