+++ version.json
@@ -1,1 +1,1 @@
-{ "version": "0.186.7" }
+{ "version": "0.186.8" }
